Renaming class Aegir\Provision\ConfigDefinition\ContextNodeDefinition to Aegir\Provision\Console\ConfigContextNodeDefinition
